Support for stopping logging
Provides new function `log_stop`, which stops logging for one or all
active log files.  This is mostly useful for tests (need to start and stop
logging multiple times to give the features a good workout) but might
be useful for some workflows (e.g., logging within a particular part
of a program).  Not yet exposed to the user.