Generalise logging approach to work with non-files
  - generalises the approach by providing `log_file` and `log_connection` as two alternative ways to start logging (note that logging to a connection will likely be significantly faster than logging by appending to a file, even with flushing after each write).
  - generalises the writing approach to allow different "types" of destination to have different writing functions.
  - fixes some extra whitespace that was being added to log messages
  - adds a `subscriptions` argument to `log_file` to help with programmatically setting up logging (see issue TODO)

This also required renaming the option `loggr_files` to `loggr_objects` (as they may not all be files).

There's a subtle change of behaviour; I kept the behaviour of "console" the same (it runs through `cat(..., file="")` but "stdout" now runs through `cat(..., file=stdout())`.  That has some weirdness with `capture.output` not working when I really feel that it should be - it's possible that something to do with the signal handling is doing something with the sinks.